Add button to remove compute node from Cluster
This is displayed within the element for the current last compute node
of each Cluster, and this node is removed when the button is clicked.

Don't currently show the button alongside every compute node as, since
compute nodes have incrementing indexes in their node names, this could
lead to holes in these indexes, and could also mean if more compute
nodes were added these could have duplicated names. There's various ways
we could improve things in future if we did want to support this, but
just allowing the user to add and then remove compute nodes as a stack
should be sufficient for now.