Copy `flight-terminal-services-client` app to adapt
Following discussion with @benarmston at
https://alces.slack.com/archives/C0BDJJXCL/p1533643896000199, and additional
IRL discussion with @jamesremuscat, it seems the best way to start developing a
new Flight client app (and keep this consistent with our other apps) is to
start with another Flight app and adapt/remove things as needed.

https://github.com/alces-software/flight-terminal-services-client is both the
most recent other Flight app started and is quite minimal, so is probably the
best choice to start from. Therefore this commit copies in the codebase as of
https://github.com/alces-software/flight-terminal-services-client/commits/f36fd7da81d813e207d005cfdebed7b5257c6ccc
to adapt.