Hack out various unneeded bits of terminal app
A lot of the copied in `flight-terminal-services-client` app isn't/won't
be needed, so hacking it out now.

Due to the nature of JavaScript, and also since some needed stuff is
mixed in with unneeded stuff, have had to be somewhat conservative when
doing this, to try to avoid things blowing up with runtime errors
anywhere at any time.