Add function to generate Nodes for PrimaryGroup
PrimaryGroups do not specify each Node within them, instead they just
specify everything required to generate every Node within the group;
this allows the group to be edited and manipulated as a whole.

This commit adds a function to generate the Nodes for a PrimaryGroup
from this specification; it also sets up `elm-test` and uses this to
test this new function.