Store PrimaryGroups outside of Clusters
Store PrimaryGroups separate to Clusters, in a UUID-indexed Dict
(actually an EveryDict as Uuid type is not `comparable` and so cannot be
put in a regular Dict), and reference these by Uuids stored in the
Clusters themselves. This should make working with groups a bit simpler
as can access the groups directly rather than needing to go via the
correct Cluster.