Switch to more comprehensible form package
I'm sure https://github.com/hecrj/composable-form is a very elegant
package, but I couldn't figure out how to customize it sufficiently to
my needs. In particular it seemed complex to customize the form markup
generated, which I want to do to display things using Bootstrap classes
and in the Bootstrap way of structuring things.

Therefore, switching to a different form package which should make this
simpler, at the expense of some possible loss of type safety (with most
of the actual customizing still to come).