Add overlay for selecting secondary group members
When member selection stage of secondary group creation is reached, add
overlay with the cluster that the secondary group is being created for
highlighted (i.e. brought in front of the overlay).

Future commits will disable other buttons for the highlighted cluster,
allow selecting groups within the cluster to be included in the
secondary group etc.