Store group ID within group record
This should make working with groups easier, as won't need to pass
around a tuple of a group with its ID if need access to both, can
instead just retrieve a group from the `model.clusterPrimaryGroups` Dict
by its ID, and then work with the retrieved group which also includes
its ID.

In order to do this needed to move ID creation from when we handle
ComputeForm submit to ComputeForm validation (as validation handles the
PrimaryGroup creation, when it is successful), and therefore need to
pass an ID to the validation function.

To make the code simpler, and avoid possible future mistakes which could
lead to duplicated IDs due to reusing a seed, each time this ID is
passed we regenerate it and save the new seed. Even though it would be
possible to only save a new seed when we actually save a new generated
ID in the model, there's no need to bother handling this as we're not
likely to run out of randomness.