Set reasonable upper bound for nodes to generate
Place upper bound for the number of nodes that can be generated in a
group, which should be high enough that nobody is likely to need more
than this. This prevents the browser locking up due to trying to
generate/handle/display 100,000 nodes if someone tried to do this for
some reason.