Validate that group names for clusters are unique
Both primary and secondary groups should have unique names within each
cluster; this commit validates that this is the case and prevents the
creation of groups with non-unique groups.

The code to do this is slightly verbose, as need to thread the current
Model and the index for the cluster the group is being created for
through several levels to the different validation functions. This works
for now but possibly we should rethink slightly how the form validation
is structured later, particularly if we later need to use the same
validation functions in contexts where there is no current cluster, or
where access to the current model is not ultimately needed.