Successfully fuzz cluster primary groups
When fuzzing Model, fuzz and save a PrimaryGroup for each fuzzed
PrimaryGroup Uuid for each Cluster.

It was quite tricky to get this consistently working both fast and
(especially) without JavaScript running out of memory on each test run.
In order to do this needed to only fuzz short lists when fuzzing lists,
and also only fuzz small ints when fuzzing `NodesSpecification`s, so
only small lists are created when using `PrimaryGroup.nodes` with a
fuzzed `NodesSpecification`; see inline comments (and code) for full
details.