Encode things as arrays when care about order
When encoding clusters, and compute groups within them, encode these as
arrays where each object has a name field, rather than as objects with
names as keys.

This is important as according to the JSON spec objects are unordered
and the order of their keys cannot be depended upon (see
https://stackoverflow.com/a/4515863/2620402). However, we do care that
the order within these fields is preserved (e.g. so when previous data
is reloaded, by encoding and decoding, everything is displayed in the
same order).