A really bad way of validating the group name
This adds validation of the group name in a bad way.  We make use of the
`validateGroupName : List PrimaryGroup -> List String -> Validation
CustomError String` function outside of an `elm-form` `Form`.

Whilst this works it has two issues:

 1. The code is kind of ugly.
 2. There is nowhere we can display error messages.

The second of these could perhaps be fixed by having `update` return a
`Cmd Msg` to set the error in the model and then we could display that
error. below the input.  However, that is likely to involve a not
insignificant amount of plumbing work which `elm-form` is designed to
handle.

A better solution then would be to change the isolated `input` element
used for the primary group name with an `elm-form`.  This would use of
`ModalForm.input` to display any errors.  The downside to this approach
is that we may end up with a large amount of types and messages to model
all of the various forms that could exist.